Conversation converter -> Conversation Filter
Changed the conversation converter structure to a filtering structure where each filter class is based on a superclass with a common method and interface which helps with scalability. Also created a FilterBuilder class which takes in a config and parseResult which determines which filters to apply based on an internal hashmap of option to filter class.