Fix existing bug/issue - only first word written to .json file
Issue was that the line taken from the txt was being split and the third element (at index 2) was being considered the message contents.
This was not the case as indext 2 would be just a single word, the remaining elements of the split contain the rest of the message content.

So, a content variable (of type String) has been introduced to correctly extract the message content from the split line.